debugged

+++ core/src/com/example/alphabetadventure/sprites/endoflevel/FireBall.java
@@ -2,12 +2,9 @@ package com.example.alphabetadventure.sprites.endoflevel;
 
 
 
-import com.badlogic.gdx.graphics.Texture;
 import com.badlogic.gdx.graphics.g2d.Animation;
-import com.badlogic.gdx.graphics.g2d.Batch;
 import com.badlogic.gdx.graphics.g2d.Sprite;
 import com.badlogic.gdx.graphics.g2d.TextureRegion;
-import com.badlogic.gdx.math.Vector2;
 import com.badlogic.gdx.physics.box2d.Body;
 import com.badlogic.gdx.physics.box2d.BodyDef;
 import com.badlogic.gdx.physics.box2d.CircleShape;
@@ -16,7 +13,6 @@ import com.badlogic.gdx.physics.box2d.World;
 import com.badlogic.gdx.utils.Array;
 import com.example.alphabetadventure.MainClass;
 import com.example.alphabetadventure.screens.PlayScreen;
-import com.example.alphabetadventure.sprites.Letter;
 
 public class FireBall extends Sprite {
 
@@ -26,10 +22,11 @@ public class FireBall extends Sprite {
     Animation fireAnimation;
     float stateTime;
     boolean destroyed;
-    boolean setToDestroy;
+    boolean fireBallSetToDestroy;
     boolean fire;
 
-    Body b1body;
+
+     Body b1body;
     public FireBall(PlayScreen screen, float x, float y, boolean fire){
         this.fire = fire;
         this.screen = screen;
@@ -37,6 +34,7 @@ public class FireBall extends Sprite {
         frames = new Array<TextureRegion>();
         for(int i = 0; i < 4; i++){
             frames.add(new TextureRegion(screen.getAtlas().findRegion("fireballs"), i * 25, 0, 25, 30));
+
         }
         fireAnimation = new Animation(0.2f, frames);
         setRegion((TextureRegion) fireAnimation.getKeyFrame(0));
@@ -52,7 +50,6 @@ public class FireBall extends Sprite {
         if(!world.isLocked())
          b1body = world.createBody(bdef);
 
-        System.out.println(world.isLocked());
         FixtureDef fdef = new FixtureDef();
         CircleShape shape = new CircleShape();
         shape.setRadius(8 / MainClass.PPM);
@@ -61,17 +58,14 @@ public class FireBall extends Sprite {
                 MainClass.GROUND_BIT |
                MainClass.PLANKS_BIT|
                 MainClass.CATAPULT_ARM_BIT|
-                MainClass.OBJECT_BIT|
                         MainClass.CATAPULT_ARM_CATCH_BIT;
 
         fdef.shape = shape;
-        fdef.restitution = 1;
-        fdef.density = 20f;
+        fdef.restitution = 1f;
+        fdef.density = 5f;
 
-        fdef.friction = 0;
-        System.out.println(b1body);
+        fdef.friction = 1f;
         b1body.createFixture(fdef).setUserData(this);
-       // b1body.setLinearVelocity(new Vector2(fire ? 2 : -2, 2.5f));
     }
 
     public void update(float dt){
@@ -79,19 +73,25 @@ public class FireBall extends Sprite {
         setRegion((TextureRegion) fireAnimation.getKeyFrame(stateTime, true));
         setPosition(b1body.getPosition().x - getWidth() / 2, b1body.getPosition().y - getHeight() / 2);
 
+        if((stateTime > 4 && fireBallSetToDestroy) && !destroyed) {
+            world.destroyBody(b1body);
+            destroyed = true;
+        }
+
+
       //  if(b1body.getLinearVelocity().y > 2f)
        //     b1body.setLinearVelocity(b1body.getLinearVelocity().x, 2f);
-       /* if((fire && b1body.getLinearVelocity().x < 0) || (!fire && b1body.getLinearVelocity().x > 0))
-            setToDestroy();*/
+        if(b1body.getPosition().x > 70 ||b1body.getLinearVelocity().x < 0)
+            fireBallSetToDestroy();
+
     }
 
 
 
 
-    public void setToDestroy(){
-        setToDestroy = true;
-        world.destroyBody(b1body);
-        destroyed = true;
+    public void fireBallSetToDestroy(){
+
+      fireBallSetToDestroy = true;
     }
 
     public boolean isDestroyed(){
@@ -99,5 +99,12 @@ public class FireBall extends Sprite {
     }
 
 
+    public void dispose() {
+
+
+    }
 }
+
+
+
